Remove the brightness slider for Blacklist apps and change behaviour
Now, the initial brightness when app-specfic brightness is enabled is the current screen brightness instead of a fixed, user specified brightness.
I think I also fixed a bug where the original screen brightness would not be applied after a non app-specific app exited onto an app-specific app, but I'm too lazy to revert and test if that bug even existed in the first place.